refactor: #43 extract hooks — useMobileState, useMobilePolling, useMobileScroll, useMobileStreaming
Shell: 795 → 576 lines (27% reduction)

Extracted hooks:
- useMobileState (179 lines): All 39 useState + derived state + refs
  Grouped by domain: core, history, review, compose, UI chrome, streaming
- useMobilePolling (192 lines): All data-fetching effects
  History loading, review packets, review files, unified sync polling,
  visibility refresh, linked owned session mapping
- useMobileScroll (73 lines): Scroll chrome, viewport tracking,
  scroll-to-bottom pinning, isWindowNearBottom
- useMobileStreaming (47 lines): WebSocket connection, hydration

Shell now contains only:
- Hook orchestration (wire the 4 hooks together)
- Action handlers (submit, enhance, copy, etc.)
- JSX render tree

Each hook is independently testable and has a single responsibility.
