+++ src/components/mobile-remote-shell.tsx
@@ -1,25 +1,18 @@
 'use client';
 import {
-  useCallback,
   useEffect,
-  useLayoutEffect,
   useMemo,
-  useRef,
-  useState,
   type CSSProperties,
 } from 'react';
 import { demoApprovals } from '@/lib/json-render/demo-specs';
 import type { ApprovalRequest } from '@/lib/json-render/demo-specs';
-import type { ReviewChangedFile, RuntimeReviewPacket } from '@/lib/fleet/types';
+import type { RuntimeReviewPacket } from '@/lib/fleet/types';
 import type {
   MobileActionRequest,
   MobileInboxSnapshot,
   MobileReviewFileResponse,
-  MobileRuntimeTailGroup,
   MobileTranscriptEntry,
-  MobileTranscriptMedia,
 } from '@/lib/mobile/types';
-import type { DraftAttachment, PendingOwnedTurn } from './mobile/types';
 import {
   agentDisplayName,
   compactLine,
@@ -40,17 +33,14 @@ const ControlsSheet = dynamic(() => import('./mobile/ControlsSheet').then((m) =>
 const CostsDashboard = dynamic(() => import('./mobile/CostsDashboard').then((m) => ({ default: m.CostsDashboard })), { ssr: false });
 const DiffOverlay = dynamic(() => import('./mobile/DiffOverlay').then((m) => ({ default: m.DiffOverlay })), { ssr: false });
 const TokenUsageSummary = dynamic(() => import('./mobile/TokenUsageSummary').then((m) => ({ default: m.TokenUsageSummary })), { ssr: false });
+
 import {
   copyTextToClipboard,
   enhancePromptDraft,
   focusSessionSurface,
   loadOwnedCorrectionDraftForSession,
-  loadOwnedReviewPacketForSession,
-  loadReviewFilePreview,
-  loadSessionHistory,
   openDiffViewerForSession,
   prepareImageAttachments,
-  refreshInboxSnapshot,
   refreshMobileSurface,
   removeImageAttachment,
   runMobileAction,
@@ -59,14 +49,13 @@ import {
   submitSteerTurn,
   updateOwnedReviewDisposition,
 } from './mobile/controller';
-import {
-  pinTranscriptToBottom,
-  startUnifiedSyncPolling,
-  trackScrollChrome,
-  trackViewportTopOffset,
-  trackVisibilityRefresh,
-} from './mobile/effects';
-import { useWebSocket } from './mobile/hooks/useWebSocket';
+
+// Extracted hooks (#43 — hooks extraction)
+import { useMobileState } from './mobile/hooks/useMobileState';
+import { useMobilePolling } from './mobile/hooks/useMobilePolling';
+import { useMobileScroll } from './mobile/hooks/useMobileScroll';
+import { useMobileStreaming } from './mobile/hooks/useMobileStreaming';
+
 export function MobileRemoteShell({
   initialSnapshot,
   initialTranscript,
@@ -78,44 +67,52 @@ export function MobileRemoteShell({
   initialReviewFile?: MobileReviewFileResponse['file'] | null;
   initialOwnedReviewPacket?: RuntimeReviewPacket | null;
 }) {
-  const [snapshot, setSnapshot] = useState<MobileInboxSnapshot>(initialSnapshot);
-  const [selectedId, setSelectedId] = useState(() => pickCurrentSession(initialSnapshot)?.id ?? '');
-  const [activeView, setActiveView] = useState<'squad' | 'chat' | 'costs'>('squad');
-  const [refreshError, setRefreshError] = useState<string | null>(null);
-  const [surfaceNote, setSurfaceNote] = useState<string | null>(null);
-  const [historyBySession, setHistoryBySession] = useState<Record<string, MobileTranscriptEntry[]>>(() => (
-    initialTranscript?.sessionKey ? { [initialTranscript.sessionKey]: initialTranscript.transcript } : {}
-  ));
-  const [historyGroupsBySession, setHistoryGroupsBySession] = useState<Record<string, MobileRuntimeTailGroup[]>>({});
-  const [historyLoading, setHistoryLoading] = useState<Record<string, boolean>>({});
-  const [historyError, setHistoryError] = useState<Record<string, string | null>>({});
-  const [reviewPacketBySession, setReviewPacketBySession] = useState<Record<string, RuntimeReviewPacket>>(() => (
-    initialOwnedReviewPacket ? { [initialOwnedReviewPacket.surfaceId]: initialOwnedReviewPacket } : {}
-  ));
-  const [reviewPacketLoadingBySession, setReviewPacketLoadingBySession] = useState<Record<string, boolean>>({});
-  const [reviewPacketErrorBySession, setReviewPacketErrorBySession] = useState<Record<string, string | null>>({});
-  const [draftBySession, setDraftBySession] = useState<Record<string, string>>({});
-  const [actionStateBySession, setActionStateBySession] = useState<Record<string, 'idle' | 'steering' | 'stopping' | 'reviewing'>>({});
-  const [actionNoteBySession, setActionNoteBySession] = useState<Record<string, string | null>>({});
-  const [draftAttachmentsBySession, setDraftAttachmentsBySession] = useState<Record<string, DraftAttachment[]>>({});
-  const [pendingOwnedTurnBySession, setPendingOwnedTurnBySession] = useState<Record<string, PendingOwnedTurn>>({});
-  const [selectedReviewFilePath, setSelectedReviewFilePath] = useState<string | null>(() => (
-    initialReviewFile?.path ?? initialOwnedReviewPacket?.changedFiles[0]?.path ?? initialSnapshot.review?.changedFiles[0]?.path ?? null
-  ));
-  const [reviewFileByPath, setReviewFileByPath] = useState<Record<string, MobileReviewFileResponse['file']>>(() => (
-    initialReviewFile ? { [initialReviewFile.path]: initialReviewFile } : {}
-  ));
-  const [reviewFileLoadingPath, setReviewFileLoadingPath] = useState<string | null>(null);
-  const [reviewFileError, setReviewFileError] = useState<string | null>(null);
-  const [diffOpen, setDiffOpen] = useState(false);
-  const [controlsOpen, setControlsOpen] = useState(false);
-  const [pendingApprovals, setPendingApprovals] = useState<ApprovalRequest[]>([]);
-  const [resolvedApprovals, setResolvedApprovals] = useState<Record<string, 'approved' | 'rejected'>>({});
-  const [enhancing, setEnhancing] = useState(false);
-  const [preEnhanceDraft, setPreEnhanceDraft] = useState<string | null>(null);
-  const [surfaceRefreshing, setSurfaceRefreshing] = useState(false);
-  const [expandedMedia, setExpandedMedia] = useState<MobileTranscriptMedia | null>(null);
-  const [scrollY, setScrollY] = useState(0);
+  // ── All state lives in useMobileState ──
+  const state = useMobileState({ initialSnapshot, initialTranscript, initialReviewFile, initialOwnedReviewPacket });
+
+  // ── Streaming + WebSocket ──
+  const { wsConnected } = useMobileStreaming(state);
+
+  // ── Data fetching + polling ──
+  const { refreshInbox, loadHistory, loadOwnedReviewPacket, loadReviewFile, reviewFiles, linkedOwnedKey } = useMobilePolling(state, wsConnected);
+
+  // ── Derived transcript data ──
+  const {
+    snapshot, selectedSession, selectedSessionKey,
+    isOwnedCodexSession, isChatSession,
+    selectedReviewPacket, selectedReviewPacketError,
+    historyBySession, historyGroupsBySession, historyLoading,
+    historyError, reviewPacketBySession, reviewPacketLoadingBySession,
+    selectedReviewFilePath, reviewFileByPath, reviewFileLoadingPath, reviewFileError,
+    diffOpen, stickyReviewFilesRef,
+    draftBySession, actionStateBySession, actionNoteBySession,
+    draftAttachmentsBySession, pendingOwnedTurnBySession,
+    enhancing, preEnhanceDraft,
+    controlsOpen, pendingApprovals, resolvedApprovals,
+    surfaceRefreshing, expandedMedia, scrollY,
+    isScrolling, headerVisible, viewportTopOffset,
+    composeFocused, waitingForResponse, hydrated,
+    expandedProject, streamingText,
+    // Setters
+    setSelectedId, setActiveView, setSurfaceNote,
+    setDraftBySession, setActionStateBySession, setActionNoteBySession,
+    setDraftAttachmentsBySession, setPendingOwnedTurnBySession,
+    setEnhancing, setPreEnhanceDraft,
+    setControlsOpen, setPendingApprovals, setResolvedApprovals,
+    setSurfaceRefreshing, setExpandedMedia,
+    setComposeFocused, setWaitingForResponse,
+    setExpandedProject, setSelectedReviewFilePath,
+    setDiffOpen, setReviewFileError,
+    setHistoryBySession,
+    // Refs
+    composeRef, fileInputRef, transcriptBottomRef,
+    stickToBottomRef, lastAssistantCountRef, seenMessageIdsRef,
+    refreshError, surfaceNote,
+    selectedId, activeView,
+    setReviewPacketBySession,
+  } = state;
+
+  // Lock body scroll when diff overlay is open
   useEffect(() => {
     if (!diffOpen) return;
     const scrollPos = window.scrollY;
@@ -131,231 +128,15 @@ export function MobileRemoteShell({
       window.scrollTo(0, scrollPos);
     };
   }, [diffOpen]);
-  const [isScrolling, setIsScrolling] = useState(false);
-  const [headerVisible, setHeaderVisible] = useState(true);
-  const [viewportTopOffset, setViewportTopOffset] = useState(0);
-  const [composeFocused, setComposeFocused] = useState(false);
-  const composeRef = useRef<HTMLTextAreaElement | null>(null);
-  const fileInputRef = useRef<HTMLInputElement | null>(null);
-  const transcriptBottomRef = useRef<HTMLDivElement | null>(null);
-  const scrollStopTimerRef = useRef<number | null>(null);
-  const headerRevealTimerRef = useRef<number | null>(null);
-  const initialBottomPinBySessionRef = useRef<Record<string, boolean>>({});
-  const stickToBottomRef = useRef(true);
-  const refreshInbox = useCallback(
-    () => refreshInboxSnapshot({ setSnapshot, setRefreshError }),
-    [],
-  );
-  const isWindowNearBottom = useCallback((threshold = 160) => {
-    if (typeof window === 'undefined') {
-      return true;
-    }
-    const scrollTop = window.scrollY || document.documentElement.scrollTop || 0;
-    const viewportBottom = scrollTop + window.innerHeight;
-    const documentHeight = Math.max(document.documentElement.scrollHeight, document.body.scrollHeight);
-    return documentHeight - viewportBottom <= threshold;
-  }, []);
-  const scrollToLatestMessage = useCallback((force = false) => {
-    if (typeof window === 'undefined') {
-      return;
-    }
-    if (!force && !stickToBottomRef.current) {
-      return;
-    }
-    transcriptBottomRef.current?.scrollIntoView({ block: 'end', behavior: 'smooth' });
-  }, []);
-  useEffect(() => {
-    return trackViewportTopOffset({ setViewportTopOffset });
-  }, []);
-  // Inbox refresh handled by unified sync polling below
-  useEffect(() => {
-    return trackScrollChrome({
-      setScrollY,
-      setIsScrolling,
-      setHeaderVisible,
-      scrollStopTimerRef,
-      headerRevealTimerRef,
-      stickToBottomRef,
-      isWindowNearBottom,
-    });
-  }, [isWindowNearBottom]);
+
+  // Sync selected session with snapshot
   useEffect(() => {
     setSelectedId((currentId) => {
-      if (currentId && snapshot.sessions.some((session) => session.id === currentId)) {
-        return currentId;
-      }
+      if (currentId && snapshot.sessions.some((session) => session.id === currentId)) return currentId;
       return pickCurrentSession(snapshot)?.id ?? '';
     });
-  }, [snapshot]);
-  const selectedSession = useMemo(
-    () => snapshot.sessions.find((session) => session.id === selectedId) ?? pickCurrentSession(snapshot),
-    [selectedId, snapshot],
-  );
-  const selectedSessionKey = selectedSession?.sessionKey;
-  const isOpenClawSession = selectedSession?.runtime === 'openclaw';
-  const isChatSession = isOpenClawSession || selectedSession?.runtime === 'codex';
-  const isOwnedCodexSession = selectedSession?.runtime === 'codex' && selectedSession?.runtimeSurface?.ownership === 'owned';
-  const selectedReviewPacket = selectedSessionKey && isOwnedCodexSession ? reviewPacketBySession[selectedSessionKey] ?? null : null;
-  const selectedReviewPacketLoading = selectedSessionKey && isOwnedCodexSession ? reviewPacketLoadingBySession[selectedSessionKey] ?? false : false;
-  const selectedReviewPacketError = selectedSessionKey && isOwnedCodexSession ? reviewPacketErrorBySession[selectedSessionKey] ?? null : null;
-  const stickyReviewFilesRef = useRef<ReviewChangedFile[]>([]);
-  const [waitingForResponse, setWaitingForResponse] = useState(false);
-  const lastAssistantCountRef = useRef(0);
-  const seenMessageIdsRef = useRef<Set<string> | null>(null);
-  const [hydrated, setHydrated] = useState(false);
-  const [streamingText, setStreamingText] = useState('');
-  const streamingTextRef = useRef('');
-  // Unified WebSocket — real-time push for inbox, history, chat deltas
-  const { connectionState: wsConnectionState } = useWebSocket({
-    selectedSessionKey,
-    setSnapshot,
-    setRefreshError,
-    setHistoryBySession,
-    setStreamingText,
-    streamingTextRef,
-  });
-  const wsConnected = wsConnectionState === 'connected';
-  useEffect(() => {
-    if (!seenMessageIdsRef.current) {
-      seenMessageIdsRef.current = new Set(transcriptEntries.map((e) => e.id));
-    }
-    setHydrated(true);
-  }, []); // eslint-disable-line react-hooks/exhaustive-deps
-  // Chat streaming is handled by useWebSocket — SSE removed (#44/#48)
-  const reviewFiles = useMemo(() => {
-    const next = isOwnedCodexSession
-      ? selectedReviewPacket?.changedFiles ?? []
-      : snapshot.review?.changedFiles ?? [];
-    if (next.length) {
-      stickyReviewFilesRef.current = next;
-      return next;
-    }
-    return stickyReviewFilesRef.current;
-  }, [isOwnedCodexSession, selectedReviewPacket, snapshot.review?.changedFiles]);
-  const loadHistory = useCallback(
-    (sessionKey: string, force = false) => loadSessionHistory({
-      sessionKey,
-      force,
-      historyBySession,
-      setHistoryLoading,
-      setHistoryBySession,
-      setHistoryGroupsBySession,
-      setHistoryError,
-    }),
-    [historyBySession],
-  );
-  const loadOwnedReviewPacket = useCallback(
-    (sessionKey: string, force = false) => loadOwnedReviewPacketForSession({
-      sessionKey,
-      force,
-      reviewPacketBySession,
-      setReviewPacketLoadingBySession,
-      setReviewPacketBySession,
-      setReviewPacketErrorBySession,
-    }),
-    [reviewPacketBySession],
-  );
-  const loadReviewFile = useCallback(
-    (reviewPath: string, force = false) => loadReviewFilePreview({
-      reviewPath,
-      force,
-      reviewFileByPath,
-      setReviewFileLoadingPath,
-      setReviewFileError,
-      setReviewFileByPath,
-    }),
-    [reviewFileByPath],
-  );
-  useEffect(() => {
-    if (!selectedSessionKey) {
-      return;
-    }
-    if (!historyBySession[selectedSessionKey]?.length && !historyLoading[selectedSessionKey]) {
-      void loadHistory(selectedSessionKey).catch(() => undefined);
-    }
-  }, [historyBySession, historyLoading, loadHistory, selectedSessionKey]);
-  useEffect(() => {
-    if (!selectedSessionKey || !selectedSessionKey.startsWith('codex-owned:')) {
-      return;
-    }
-    if (!reviewPacketBySession[selectedSessionKey] && !reviewPacketLoadingBySession[selectedSessionKey]) {
-      void loadOwnedReviewPacket(selectedSessionKey).catch(() => undefined);
-    }
-  }, [loadOwnedReviewPacket, reviewPacketBySession, reviewPacketLoadingBySession, selectedSessionKey]);
-  useEffect(() => {
-    if (!reviewFiles.length) {
-      setSelectedReviewFilePath(null);
-      setReviewFileError(null);
-      return;
-    }
-    if (selectedReviewFilePath && reviewFiles.some((file) => file.path === selectedReviewFilePath)) {
-      return;
-    }
-    const nextPath = reviewFiles[0]?.path ?? null;
-    setSelectedReviewFilePath(nextPath);
-    if (nextPath) {
-      void loadReviewFile(nextPath).catch(() => undefined);
-    }
-  }, [loadReviewFile, reviewFiles, selectedReviewFilePath]);
-  const documentVisibleRef = useRef(true);
-  useEffect(() => {
-    return trackVisibilityRefresh({
-      documentVisibleRef,
-      selectedSessionKey,
-      loadHistory,
-      refreshInbox,
-    });
-  }, [loadHistory, refreshInbox, selectedSessionKey]);
-  const linkedOwnedKey = useMemo(() => {
-    if (!selectedSession || selectedSession.runtime !== 'codex' || selectedSession.runtimeSurface?.ownership !== 'discovered') return null;
-    const cwd = selectedSession.runtimeSurface?.cwd ?? selectedSession.workspace ?? '';
-    const owned = snapshot.sessions.find((s) =>
-      s.runtime === 'codex' &&
-      s.runtimeSurface?.ownership === 'owned' &&
-      (s.runtimeSurface?.cwd === cwd || s.workspace === cwd),
-    );
-    return owned?.sessionKey ?? null;
-  }, [selectedSession, snapshot.sessions]);
-  // Initial linked history load
-  useEffect(() => {
-    if (linkedOwnedKey && !historyBySession[linkedOwnedKey] && !historyLoading[linkedOwnedKey]) {
-      void loadHistory(linkedOwnedKey, true).catch(() => undefined);
-    }
-  }, [linkedOwnedKey, historyBySession, historyLoading, loadHistory]);
-  // Unified sync polling — replaces separate session + linked + review file polling (5 requests → 1)
-  useEffect(() => {
-    return startUnifiedSyncPolling({
-      selectedSessionKey,
-      selectedSession,
-      linkedOwnedKey,
-      pendingOwnedTurnBySession,
-      actionStateBySession,
-      waitingForResponse,
-      diffOpen,
-      selectedReviewFilePath,
-      documentVisibleRef,
-      historyBySession,
-      wsConnected,
-      setSnapshot,
-      setRefreshError,
-      setHistoryBySession,
-      setHistoryGroupsBySession,
-      setReviewFileByPath,
-      loadOwnedReviewPacket,
-    });
-  }, [
-    actionStateBySession,
-    diffOpen,
-    historyBySession,
-    linkedOwnedKey,
-    loadOwnedReviewPacket,
-    pendingOwnedTurnBySession,
-    selectedReviewFilePath,
-    selectedSession,
-    selectedSessionKey,
-    waitingForResponse,
-    wsConnected,
-  ]);
+  }, [snapshot, setSelectedId]);
+
   const effectiveHistoryKey = linkedOwnedKey && historyBySession[linkedOwnedKey]?.length ? linkedOwnedKey : selectedSessionKey;
   const discoveredEntries = selectedSessionKey ? historyBySession[selectedSessionKey] ?? [] : [];
   const ownedEntries = linkedOwnedKey ? historyBySession[linkedOwnedKey] ?? [] : [];
@@ -370,36 +151,28 @@ export function MobileRemoteShell({
   const transcriptAttachments = selectedSessionKey ? draftAttachmentsBySession[selectedSessionKey] ?? [] : [];
   const pendingOwnedTurn = selectedSessionKey ? pendingOwnedTurnBySession[selectedSessionKey] ?? null : null;
   const transcriptActionState = selectedSessionKey ? actionStateBySession[selectedSessionKey] ?? 'idle' : 'idle';
+  const transcriptActionNote = selectedSessionKey ? actionNoteBySession[selectedSessionKey] ?? null : null;
+  const selectedReviewFile = selectedReviewFilePath ? reviewFileByPath[selectedReviewFilePath] : undefined;
+
+  // Track assistant count for response detection
   const assistantCount = transcriptEntries.filter((e) => e.role === 'assistant').length;
   useEffect(() => {
     if (waitingForResponse && assistantCount > lastAssistantCountRef.current) {
       setWaitingForResponse(false);
     }
-  }, [waitingForResponse, assistantCount]);
-  const transcriptActionNote = selectedSessionKey ? actionNoteBySession[selectedSessionKey] ?? null : null;
+  }, [waitingForResponse, assistantCount, lastAssistantCountRef, setWaitingForResponse]);
+
   const latestTranscriptMarker = transcriptEntries[transcriptEntries.length - 1]?.id ?? 'empty';
   const scrollMarker = pendingOwnedTurn ? `${latestTranscriptMarker}:${pendingOwnedTurn.id}` : latestTranscriptMarker;
-  const selectedReviewFile = selectedReviewFilePath ? reviewFileByPath[selectedReviewFilePath] : undefined;
-  const [expandedProject, setExpandedProject] = useState<string | null>(null);
-  const sessionSwitcher = snapshot.sessions.slice(0, 5);
-  const headerProgress = Math.min(scrollY / 88, 1);
-  const isHeaderCompact = headerProgress > 0.12;
-  const isComposerPrimed = isChatSession && (composeFocused || transcriptAttachments.length > 0);
-  const dockMotionProgress = !isComposerPrimed && isScrolling ? 1 : 0;
-  const dockFadeProgress = dockMotionProgress;
-  const ownedAvailability = selectedSession?.runtimeSurface?.lifecycle?.availability;
-  const ownedReviewDisposition = selectedReviewPacket?.reviewDisposition;
-  const ownedQueuedTurn = Boolean(pendingOwnedTurn) || transcriptActionState === 'steering';
-  const canResumeOwnedCodex = Boolean(isOwnedCodexSession && selectedSession?.runtimeSurface?.capabilities.sendInput && !ownedQueuedTurn);
-  const canInterruptOwnedCodex = Boolean(isOwnedCodexSession && selectedSession?.runtimeSurface?.capabilities.interrupt);
+
+  // ── Scroll management ──
+  const { scrollToLatestMessage } = useMobileScroll(state, transcriptEntries, transcriptGroups, pendingOwnedTurn, scrollMarker);
+
+  // ── Owned Codex turn lifecycle ──
   useEffect(() => {
-    if (!selectedSessionKey?.startsWith('codex-owned:')) {
-      return;
-    }
+    if (!selectedSessionKey?.startsWith('codex-owned:')) return;
     const pendingTurn = pendingOwnedTurnBySession[selectedSessionKey];
-    if (!pendingTurn) {
-      return;
-    }
+    if (!pendingTurn) return;
     const sessionGroups = historyGroupsBySession[selectedSessionKey] ?? [];
     const matchingGroup = sessionGroups.find((group) => {
       const promptMatches = group.prompt.trim() === pendingTurn.prompt.trim();
@@ -411,18 +184,28 @@ export function MobileRemoteShell({
       && !selectedSession?.runtimeSurface?.capabilities.interrupt
       && transcriptActionState === 'idle',
     );
-    if (!matchingGroup && !runSettledAgain) {
-      return;
-    }
+    if (!matchingGroup && !runSettledAgain) return;
     setPendingOwnedTurnBySession((current) => {
-      if (!current[selectedSessionKey]) {
-        return current;
-      }
+      if (!current[selectedSessionKey]) return current;
       const next = { ...current };
       delete next[selectedSessionKey];
       return next;
     });
-  }, [historyGroupsBySession, pendingOwnedTurnBySession, selectedSession?.runtimeSurface?.capabilities.interrupt, selectedSession?.runtimeSurface?.capabilities.sendInput, selectedSessionKey, transcriptActionState]);
+  }, [historyGroupsBySession, pendingOwnedTurnBySession, selectedSession?.runtimeSurface?.capabilities.interrupt, selectedSession?.runtimeSurface?.capabilities.sendInput, selectedSessionKey, transcriptActionState, setPendingOwnedTurnBySession]);
+
+  // ── UI layout values ──
+  const sessionSwitcher = snapshot.sessions.slice(0, 5);
+  const headerProgress = Math.min(scrollY / 88, 1);
+  const isHeaderCompact = headerProgress > 0.12;
+  const isComposerPrimed = isChatSession && (composeFocused || transcriptAttachments.length > 0);
+  const dockMotionProgress = !isComposerPrimed && isScrolling ? 1 : 0;
+  const dockFadeProgress = dockMotionProgress;
+  const ownedAvailability = selectedSession?.runtimeSurface?.lifecycle?.availability;
+  const ownedReviewDisposition = selectedReviewPacket?.reviewDisposition;
+  const ownedQueuedTurn = Boolean(pendingOwnedTurn) || transcriptActionState === 'steering';
+  const canResumeOwnedCodex = Boolean(isOwnedCodexSession && selectedSession?.runtimeSurface?.capabilities.sendInput && !ownedQueuedTurn);
+  const canInterruptOwnedCodex = Boolean(isOwnedCodexSession && selectedSession?.runtimeSurface?.capabilities.interrupt);
+
   const shellStyle = {
     '--remodex-header-progress': headerProgress.toFixed(3),
     '--remodex-dock-fade-progress': dockFadeProgress.toFixed(3),
@@ -430,26 +213,12 @@ export function MobileRemoteShell({
     '--remodex-compose-active': isComposerPrimed ? '1' : '0',
     '--remodex-viewport-top-offset': `${viewportTopOffset}px`,
   } as CSSProperties;
-  useLayoutEffect(() => {
-    return pinTranscriptToBottom({
-      selectedSessionKey,
-      transcriptEntries,
-      transcriptGroups,
-      pendingOwnedTurn,
-      initialBottomPinBySessionRef,
-      transcriptBottomRef,
-      stickToBottomRef,
-    });
-  }, [pendingOwnedTurn, scrollMarker, selectedSessionKey, transcriptEntries, transcriptGroups]);
-  async function handleAttachmentSelection(files: FileList | null) {
-    await prepareImageAttachments({ selectedSessionKey, files, isChatSession, setSurfaceNote, setDraftAttachmentsBySession, composeRef });
-  }
-  function removeDraftAttachment(sessionKey: string, attachmentId: string) {
-    removeImageAttachment({ sessionKey, attachmentId, setDraftAttachmentsBySession });
-  }
+
+  // ── Action handlers ──
   async function runAction(payload: MobileActionRequest) {
     return await runMobileAction({ payload, setActionStateBySession, setActionNoteBySession, refreshInbox, loadHistory, loadOwnedReviewPacket });
   }
+
   function playSendClick() {
     try {
       const ctx = new (window.AudioContext || (window as unknown as { webkitAudioContext: typeof AudioContext }).webkitAudioContext)();
@@ -465,14 +234,25 @@ export function MobileRemoteShell({
       osc.stop(ctx.currentTime + 0.06);
     } catch { /* audio not available */ }
   }
+
+  async function handleAttachmentSelection(files: FileList | null) {
+    await prepareImageAttachments({ selectedSessionKey, files, isChatSession, setSurfaceNote, setDraftAttachmentsBySession, composeRef });
+  }
+
+  function removeDraftAttachment(sessionKey: string, attachmentId: string) {
+    removeImageAttachment({ sessionKey, attachmentId, setDraftAttachmentsBySession });
+  }
+
   async function handleEnhancePrompt() {
     await enhancePromptDraft({ selectedSessionKey, enhancing, draftBySession, setEnhancing, setPreEnhanceDraft, setDraftBySession, setSurfaceNote });
   }
+
   function handleUndoEnhance() {
     if (!selectedSessionKey || preEnhanceDraft === null) return;
     setDraftBySession((current) => ({ ...current, [selectedSessionKey]: preEnhanceDraft }));
     setPreEnhanceDraft(null);
   }
+
   async function handleSteerSubmit(sessionKey: string) {
     await submitSteerTurn({
       sessionKey,
@@ -496,9 +276,11 @@ export function MobileRemoteShell({
       playSendClick,
     });
   }
+
   function handleLoadOwnedCorrectionDraft(sessionKey: string) {
     loadOwnedCorrectionDraftForSession({ sessionKey, reviewPacketBySession, setDraftBySession, setActionNoteBySession, composeRef });
   }
+
   async function handleOwnedResumeSubmit(sessionKey: string) {
     await submitOwnedResumeTurn({
       sessionKey,
@@ -512,6 +294,7 @@ export function MobileRemoteShell({
       playSendClick,
     });
   }
+
   async function handleOwnedReviewDisposition(action: 'watch' | 'resolve', sessionKey: string) {
     await updateOwnedReviewDisposition({
       action,
@@ -524,9 +307,11 @@ export function MobileRemoteShell({
       loadOwnedReviewPacket,
     });
   }
+
   function handleCopy(text: string) {
     copyTextToClipboard({ text, setSurfaceNote });
   }
+
   async function handleSurfaceRefresh() {
     setSurfaceRefreshing(true);
     try {
@@ -543,6 +328,7 @@ export function MobileRemoteShell({
       setSurfaceRefreshing(false);
     }
   }
+
   function handleSessionFocus(sessionId: string) {
     focusSessionSurface({
       sessionId,
@@ -559,9 +345,11 @@ export function MobileRemoteShell({
       loadReviewFile,
     });
   }
+
   async function handleStopActiveRun() {
     await stopActiveRunFromSurface({ selectedSessionKey, isChatSession, canInterruptOwnedCodex, isOwnedCodexSession, runAction, setSurfaceNote, setControlsOpen });
   }
+
   function openDiffViewer() {
     openDiffViewerForSession({
       reviewFiles,
@@ -574,35 +362,35 @@ export function MobileRemoteShell({
       loadReviewFile,
     });
   }
+
   function handleReviewFileFocus(reviewPath: string) {
     setSelectedReviewFilePath(reviewPath);
     void loadReviewFile(reviewPath).catch(() => undefined);
   }
+
   function handleApprovalDecision(approval: ApprovalRequest, resolution: 'approved' | 'rejected') {
     setResolvedApprovals((current) => ({ ...current, [approval.id]: resolution }));
     setSurfaceNote(`${resolution === 'approved' ? '✅ Approved' : '❌ Rejected'}: ${approval.title}`);
     window.setTimeout(() => setPendingApprovals((current) => current.filter((item) => item.id !== approval.id)), 1500);
   }
-  function handleApprovalApprove(approval: ApprovalRequest) {
-    handleApprovalDecision(approval, 'approved');
-  }
-  function handleApprovalReject(approval: ApprovalRequest) {
-    handleApprovalDecision(approval, 'rejected');
-  }
+
   function handleToggleApprovals() {
     setPendingApprovals((current) => (current.length > 0 ? [] : [...demoApprovals]));
     setResolvedApprovals({});
     setControlsOpen(false);
   }
+
   function handleCopySelectedSessionKey() {
     if (!selectedSessionKey) return;
     handleCopy(selectedSessionKey);
     setControlsOpen(false);
   }
+
   function handleControlsRefresh() {
     void handleSurfaceRefresh();
     setControlsOpen(false);
   }
+
   function handleDiffRefresh() {
     if (selectedReviewFilePath) {
       void loadReviewFile(selectedReviewFilePath, true);
@@ -610,8 +398,10 @@ export function MobileRemoteShell({
     }
     void handleSurfaceRefresh();
   }
+
   const withSelectedSession = <Args extends unknown[]>(fn: (sessionKey: string, ...args: Args) => void | Promise<void>) =>
     (...args: Args): void | Promise<void> => (selectedSessionKey ? fn(selectedSessionKey, ...args) : undefined);
+
   const composeBarHandlers = {
     onSend: withSelectedSession(handleSteerSubmit),
     onOwnedResume: withSelectedSession(handleOwnedResumeSubmit),
@@ -629,6 +419,8 @@ export function MobileRemoteShell({
     onDraftChange: withSelectedSession((sessionKey, value: string) => setDraftBySession((current) => ({ ...current, [sessionKey]: value }))),
     onFocusChange: setComposeFocused,
   };
+
+  // ── Render ──
   return (
     <div className="mobile-wrap remodex-mobile-page" style={shellStyle} suppressHydrationWarning>
       <div className="remodex-phone-shell">
@@ -703,8 +495,8 @@ export function MobileRemoteShell({
           <ApprovalStack
             pendingApprovals={pendingApprovals}
             resolvedApprovals={resolvedApprovals}
-            onApprove={handleApprovalApprove}
-            onReject={handleApprovalReject}
+            onApprove={(a) => handleApprovalDecision(a, 'approved')}
+            onReject={(a) => handleApprovalDecision(a, 'rejected')}
           />
           <div ref={transcriptBottomRef} className="remodex-scroll-anchor" aria-hidden="true" />
         </div>
